% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{icd10cm_codes_2019}
\alias{icd10cm_codes_2019}
\title{icd-10-cm list and descriptions original.}
\format{Data frame}
\source{
\url{https://www.cdc.gov/nchs/icd/icd10cm.htm}
}
\usage{
icd10cm_codes_2019
}
\description{
Dataset of 143,865 rows and 2 variables.
}
\examples{
tail(icd10cm_codes_2019)
}
\keyword{datasets}
