% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{icd10cm_codes_regex}
\alias{icd10cm_codes_regex}
\title{icd-10-cm list and descriptions original.}
\format{
Data frame
}
\source{
\url{https://www.cms.gov/medicare/icd-10/2021-icd-10-cm}
}
\usage{
icd10cm_codes_regex
}
\description{
Dataset of 72,616 rows and 3 variables.
}
\examples{
tail(icd10cm_codes_regex)
}
\keyword{datasets}
