% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{icd10cm_injury_matrix}
\alias{icd10cm_injury_matrix}
\title{The complete icd10cm injury matrix.}
\format{
Data frame
}
\source{
\url{ftp://ftp.cdc.gov/pub/Health_Statistics/NCHS/injury/tools/}
}
\usage{
icd10cm_injury_matrix
}
\description{
Dataset of 3,655 rows and 5 variables.
formatted from the original
}
\examples{
library(dplyr)
sample_n(icd10cm_injury_matrix, 10)
}
\keyword{datasets}
