% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{icd10cm_mdc_drg_regex}
\alias{icd10cm_mdc_drg_regex}
\title{Major Diagnostic Categories (MDC), diagnosis-related group (DRG), and ICD-10-CM .}
\format{
Data frame
}
\source{
\url{https://www.cms.gov/icd10m/version39-fullcode-cms/fullcode_cms/P0001.html}
}
\usage{
icd10cm_mdc_drg_regex
}
\description{
Dataset of 3594355 rows and 9 variables.
}
\examples{
tail(icd10cm_mdc_drg_regex)
}
\keyword{datasets}
