% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_age_group10.R
\name{icd_age_group10}
\alias{icd_age_group10}
\title{Cut age into ten age groups}
\usage{
icd_age_group10(data, age)
}
\arguments{
\item{data}{input data}

\item{age}{age as continuous variable}
}
\value{
return an additional age_group10 fields of ten age intervals
}
\description{
Cut age into ten age groups
}
\examples{
library(dplyr)
library(classInt)
set.seed(10)
dat <- tibble(age = sample(c(1:100, NA), 110, replace = TRUE))
dat <- dat \%>\% icd_age_group10(age = age)
dat

}
