% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_clean_mech_names.r
\name{icd_clean_mech_names}
\alias{icd_clean_mech_names}
\title{clean the mechanism categories replacing space and other symbols
 with _.}
\usage{
icd_clean_mech_names(x)
}
\arguments{
\item{x}{input variable}
}
\value{
return cleaner mechanism categories
}
\description{
clean the mechanism categories replacing space and other symbols
 with _.
}
\examples{
# library(tidyverse)
icd_clean_mech_names(c("Natural/Environmental, Other"))


}
