% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_utils.R
\name{icd_create_indicator}
\alias{icd_create_indicator}
\title{Create a new indicator based on pattern in the argument expr}
\usage{
icd_create_indicator(data, new_name, expr, colvec, ignore.case = T, perl = T)
}
\arguments{
\item{data}{input data}

\item{new_name}{proposed name for the indicator}

\item{expr}{regular expression describing the pattern of interest}

\item{colvec}{indices or names of variables (where are the pattern) without quotes}

\item{ignore.case}{logical}

\item{perl}{logical}
}
\value{
new indicator matching the pattern described in the regular expression
}
\description{
Create a new indicator based on pattern in the argument expr
}
\examples{

library(dplyr)
library(useicd10cm)
icd10cm_data150 \%>\%
  icd_create_indicator(new_name = "hero", expr = "T401.[1-4]", colvec = c(2:6))) \%>\%
  count(hero)
}
