% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_drug_opioid.R
\name{icd_drug_opioid}
\alias{icd_drug_opioid}
\title{Find drug types from ICD-10-CM.}
\usage{
icd_drug_opioid(data, diag_ecode_col)
}
\arguments{
\item{data}{input data}

\item{diag_ecode_col}{column indices}
}
\value{
any_drug, any_opioid, non_heroin_opioid,
  heroin, stimulant
}
\description{
Find any drug, selected opioids and stimulants
}
\examples{
library(tidyverse)
icd10cm_data150 \%>\% icd_drug_opioid(diag_ecode_col = c(2:6)) \%>\% sample_n(10)

}
