% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_drug_opioid_ish.R
\name{icd_drug_opioid_isf}
\alias{icd_drug_opioid_isf}
\title{Find intentional self-harm drug types from ICD-10-CM.}
\usage{
icd_drug_opioid_isf(data, diag_ecode_col)
}
\arguments{
\item{data}{input data}

\item{diag_ecode_col}{column indices}
}
\value{
any_drug_isf, any_opioid_isf, non_heroin_opioid_isf,
  heroin_isf, stimulant_isf, cocaine_isf, stimulant_not_cocaine_isf
}
\description{
Find any drug, selected opioids and stimulants
}
\examples{

library(dplyr)
library(purrr)
icd10cm_data150 \%>\%
  icd_drug_opioid(diag_ecode_col = c(2:6)) \%>\%
  sample_n(10)
}
