% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_drug_opioid_ish.R
\name{icd_drug_opioid_ish}
\alias{icd_drug_opioid_ish}
\title{Find intentional self-harm drug types from ICD-10-CM.}
\usage{
icd_drug_opioid_ish(data, diag_ecode_col)
}
\arguments{
\item{data}{input data}

\item{diag_ecode_col}{column indices}
}
\value{
any_drug_ish, any_opioid_ish, non_heroin_opioid_ish,
  heroin_ish, stimulant_ish, cocaine_ish, stimulant_not_cocaine_ish
}
\description{
Find any drug, selected opioids and stimulants with intentional self-harm
}
\examples{

library(dplyr)
library(purrr)
icd10cm_data150 \%>\%
  icd_drug_opioid_ish(diag_ecode_col = c(2:6)) \%>\%
  sample_n(10)
}
