% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10cm_drug.R
\name{icd_drug_opioid_uu}
\alias{icd_drug_opioid_uu}
\title{Find unintetional and undetermined drug types from ICD-10-CM.}
\usage{
icd_drug_opioid_uu(data, diag_ecode_col)
}
\arguments{
\item{data}{input data}

\item{diag_ecode_col}{column indices}
}
\value{
any_drug_uu, any_opioid_uu, non_heroin_opioid_uu,
  heroin_uu, stimulant_uu, cocaine_uu, stimulant_not_cocaine_uu
}
\description{
Find any drug, selected opioids and stimulants
}
\examples{

library(dplyr)
library(purrr)
icd10cm_data150 \%>\%
  icd_drug_opioid(diag_ecode_col = c(2:6)) \%>\%
  sample_n(10)
}
