% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_utils.R
\name{icd_first_match}
\alias{icd_first_match}
\title{a row operation that will form a vector of the first match of a pattern.}
\usage{
icd_first_match(data, new_name, colvec, pattern)
}
\arguments{
\item{data}{input data}

\item{new_name}{proposed name for the new variable}

\item{colvec}{selected columns to match}

\item{pattern}{the pattern to match}
}
\value{
return the vector of the matched characters with NA for a no match
}
\description{
a row operation that will form a vector of the first match of a pattern.
}
\examples{

dat <- data.frame(x1 = letters[1:3], x2 = c("d", "a", "e"))
library(dplyr)
library(purrr)
dat \%>\% icd_first_match(new_name = "x3", colvec = c(1:2), pattern = "a")
}
