% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_first_valid.R
\name{icd_first_valid}
\alias{icd_first_valid}
\title{a row operation that will form a vector of the first match of a pattern.}
\usage{
icd_first_valid(data, colvec, pattern)
}
\arguments{
\item{data}{input data}

\item{colvec}{selected columns to match}

\item{pattern}{the pattern to match}
}
\value{
return the vector of the matched characters with NA for a no match
}
\description{
a row operation that will form a vector of the first match of a pattern.
}
\examples{
dat <- tibble(x1 = letters[1:3], x2 = c("d", "a", "e"))
dat \%>\% mutate(x3 = icd_first_valid(., colvec = c(1:2), pattern = "a"))
library(tidyverse)
icd_make_regex1(c("X45x2", "Y65xx"))


}
