% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_utils.R
\name{icd_first_valid_index}
\alias{icd_first_valid_index}
\title{Row operation that creates a vector of indices the first match of a pattern}
\usage{
icd_first_valid_index(data, colvec, pattern)
}
\arguments{
\item{data}{input data}

\item{colvec}{selected columns to match}

\item{pattern}{the pattern to match}
}
\value{
return the vector of the indices of the matches with 0 for no match
}
\description{
Row operation that creates a vector of indices the first match of a pattern
}
\examples{

dat <- data.frame(x1 = letters[1:3], x2 = c("d", "a", "e"))
library(dplyr)
library(purrr)
dat \%>\% mutate(x3 = icd_first_valid_index(., colvec = c(1:2), pattern = "a"))

}
