% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_injury.R
\name{icd_injury}
\alias{icd_injury}
\title{Find drug types from ICD-10-CM.}
\usage{
icd_injury(data, diag_ecode_col)
}
\arguments{
\item{data}{input data}

\item{diag_ecode_col}{column indices}
}
\value{
any_drug, any_opioid, non_heroin_opioid,
  heroin, stimulant, cocaine, stimulant_not_cocaine
}
\description{
Classify injuries from ICD-10-CM codes
}
\examples{

library(dplyr)
library(purrr)
icd10cm_data150 \%>\%
  icd_drug_opioid(diag_ecode_col = c(2:6)) \%>\%
  sample_n(10)
}
