% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10cm_injury.R
\name{icd_injury_regex}
\alias{icd_injury_regex}
\title{Selected injury ICD 10 CM regular expressions}
\usage{
icd_injury_regex()
}
\value{
Return objects of regular expressions in the global environment
}
\description{
Selected injury ICD 10 CM regular expressions
}
\examples{

icd_injury_regex()
grep("^icd10cm.+_$", objects(), value = TRUE)

}
