% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_make_regex1.R
\name{icd_make_regex1}
\alias{icd_make_regex1}
\title{transform the ICD-10-CM codes into an operational regular expression.}
\usage{
icd_make_regex1(x)
}
\arguments{
\item{x}{input variable}
}
\value{
return the icd10cm as regular expressions with all the "x" not at the beginning of the icd10cm code replaced by '.' to match any character.
}
\description{
transform the ICD-10-CM codes into an operational regular expression.
}
\examples{
to be added
library(tidyverse)
icd_make_regex1(c("X45x2", "Y65xx"))


}
