% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10cm_injury.R
\name{icd_matched_mechanism}
\alias{icd_matched_mechanism}
\title{Add matched mechanism fields for injury ICD-10-CM.}
\usage{
icd_matched_mechanism(data, inj_col)
}
\arguments{
\item{data}{input data}

\item{inj_col}{ecode and diagnosis column indices}
}
\value{
return the input with additional variables (32 mechanisms)
}
\description{
, reference
}
\examples{
library(tidyverse)
library(fuzzyjoin)
dat <- data.frame(
  d1 = c("T63023", "X92821", "X99100", "T360x"),
  d2 = c("T65823", "Y030x0", "T17200", "V0100x")
)

dat \%>\% icd_matched_mechanism(inj_col = c(1, 2))

}
