% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_new_var.R
\name{icd_new_var}
\alias{icd_new_var}
\title{Create a new variable without mutate based on pattern in the argument expr}
\usage{
icd_new_var(data, expr, colvec, var_name = "new_var", ignore.case = T,
  perl = T)
}
\arguments{
\item{data}{input data}

\item{expr}{regular expression describing the pattern of interest}

\item{colvec}{indices of variables of interest}

\item{var_name}{the proposed name of the new variable. The default is "new_var"}

\item{ignore.case}{logical}

\item{perl}{logical}
}
\value{
new variable matching the pattern described in the regular expression
}
\description{
Create a new variable without mutate based on pattern in the argument expr
}
\examples{

library(dplyr)
library(purrr)
set.seed(2)
icd10cm_data150 \%>\%
icd_new_var(expr = "T401.[1-4]", colvec = c(2:6), var_name = "hero") \%>\%
sample_n(10)

}
