% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_od_definitions.r
\name{icd_od_definitions}
\alias{icd_od_definitions}
\title{Selected OD ICD 10 CM definitions}
\usage{
icd_od_definitions()
}
\value{
Return seven definitions and their regular expressions
}
\description{
Selected OD ICD 10 CM definitions
}
\examples{

library(crayon)
useicd10cm::icd_od_definitions()

}
