% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10cm_injury.R
\name{icd_select_intent}
\alias{icd_select_intent}
\title{Add Selected Intents}
\usage{
icd_select_intent(data, inj_col, ...)
}
\arguments{
\item{data}{input data}

\item{inj_col}{ecode and diagnosis column indices}

\item{...}{keyword list}
}
\value{
return the input with additional intent variables
}
\description{
Add Selected Intents
}
\examples{
library(dplyr)
library(purrr)
dat <- data.frame(
  d1 = c("T63023", "X92821", "X99100", "T360x"),
  d2 = c("T65823", "Y030x0", "T17200", "V0100x")
)

dat \%>\% icd_select_intent(inj_col = c(1, 2), "untintentional", "assault")

}
