% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10cm_injury.R
\name{icd_select_mechanism}
\alias{icd_select_mechanism}
\title{Add Selected Mechanisms}
\usage{
icd_select_mechanism(data, inj_col, ...)
}
\arguments{
\item{data}{input data}

\item{inj_col}{ecode and diagnosis column indices}

\item{...}{keyword list}
}
\value{
return the input with additional mechanism variables
}
\description{
Add Selected Mechanisms
}
\examples{
library(dplyr)
dat <- data.frame(
  d1 = c("T63023", "X92821", "X99100", "T360x"),
  d2 = c("T65823", "Y030x0", "T17200", "V0100x")
)

dat \%>\% icd_select_mechanism(inj_col = c(1, 2), "fall", "firearm") # by default all mechanisms

}
