% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10cm_injury.R
\name{icd_svip}
\alias{icd_svip}
\title{Find core svipp indicators from ICD-10-CM.}
\source{
\url{https://www.cdc.gov/injury/pdfs/2019_state_injury_indicator_instructions-508.pdf}
}
\usage{
icd_svip(data, diag_ecode_col)
}
\arguments{
\item{data}{input data}

\item{diag_ecode_col}{column indices}
}
\value{
selected core svipp nonfatal injury indicators
}
\description{
Find nonfatal injury indicators
}
\examples{

library(dplyr)
library(purrr)
icd10cm_data150 \%>\%
  icd_svip(diag_ecode_col = c(2:6)) \%>\%
  sample_n(10)
}
