% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10cm_injury.R
\name{icd_svip_injury}
\alias{icd_svip_injury}
\title{Find core svipp valid injury cases from ICD-10-CM.}
\source{
\url{https://www.cdc.gov/injury/pdfs/2019_state_injury_indicator_instructions-508.pdf}
}
\usage{
icd_svip_injury(data, diag_col)
}
\arguments{
\item{data}{input data}

\item{diag_col}{principal diagnosis code column}
}
\value{
core svipp nonfatal injury cases
}
\description{
Find nonfatal injury cases
}
\examples{

library(dplyr)
library(purrr)
icd10cm_data150 \%>\%
  icd_svip_injury(diag_col = c(2:6)) \%>\%
  sample_n(10)
}
