% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useicd10cm-package.R
\docType{package}
\name{useicd10cm-package}
\alias{useicd10cm}
\alias{useicd10cm-package}
\title{useicd10cm: functions and data for the use of ICD-10-CM codes}
\description{
Functions and data for the implementation of ICD-10-CM in injury epidemiology including overdose
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/epinotes/useicd10cm}
  \item Report bugs at \url{https://github.com/epinotes/useicd10cm/issues}
}

}
\keyword{internal}
