% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{wa_fips_codes}
\alias{wa_fips_codes}
\title{County FIPS codes for WA State.}
\format{Data frame}
\source{
US Census Bureau Through the the R package"tidycensus"
}
\usage{
wa_fips_codes
}
\description{
Dataset of 39 rows and 7 variables.
}
\examples{
head(wa_fips_codes)

}
\keyword{datasets}
