% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{icd10cm_codes_2019}
\alias{icd10cm_codes_2019}
\title{icd-10-cm list and descriptions original.}
\format{Data frame}
\source{
\url{https://www.cms.gov/Medicare/Coding/ICD10/2019-ICD-10-CM}
}
\usage{
icd10cm_codes_2019
}
\description{
Dataset of 71,932 rows and 3 variables.
}
\examples{
tail(icd10cm_codes_2019)
}
\keyword{datasets}
