% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_new_diag.R
\name{icd_new_diag}
\alias{icd_new_diag}
\title{Create a new variable based on pattern in the argument expr}
\usage{
icd_new_diag(data, expr, colvec, ignore.case = T, perl = T)
}
\arguments{
\item{ignore.case}{logical}

\item{perl}{logical}

\item{data:}{input data}

\item{expr:}{regular expression describing the pattern of interest}

\item{colvec:}{indices of variables of interest}
}
\value{
new variable matching the pattern described in the regular expression
}
\description{
Create a new variable based on pattern in the argument expr
}
\examples{

library(dplyr)
library(purrr)
icd10cm_data150 \%>\%
  mutate(hero = icd_new_diag(., expr = "T401.[1-4]", colvec = c(2:6))) \%>\%
  count(hero)
}
